stark: proof serialization and the attestation verification core
serialize/deserialize give a proof a canonical byte form for a trailer:
the decoder is total over hostile bytes, allocates only what the input
can back, and rejects a non-canonical element or trailing bytes, so the
round trip is exact.

attest verifies a context-bound membership proof against the kernel's
own trusted root; attest_trailer parses a whole trailer and verifies
it, with the round and query counts fixed kernel-side so a prover
cannot weaken the low-degree test.

Tests over a real Poseidon tree: an enrolled capsule is accepted; a
different capsule, a wrong root, a tampered proof, a malformed path,
and every truncation are rejected without a panic.
